+++ dejacode/__init__.py
@@ -70,8 +70,7 @@ __version__ = get_version(VERSION)
 
 # Turn off the warnings for the following modules.
 warnings.filterwarnings("ignore", module="cyclonedx")
-# TODO:
-warnings.filterwarnings("ignore", module="crontab")
+warnings.filterwarnings("ignore", category=FutureWarning, module="rq_scheduler.utils")
 
 
 def command_line():